+++ backend/api/__init__.py
@@ -1,0 +1,1 @@
+# Empty file to make dir into a module
