+++ backend/entities/album_entity.py
@@ -1,0 +1,10 @@
+from .entity_base import EntityBase
+from sqlalchemy import DateTime, Enum, ForeignKey, Integer, String
+from sqlalchemy.orm import Mapped, mapped_column, relationship
+
+
+class AlbumEntity(EntityBase):
+    __tablename__ = "album"
+
+    id: Mapped[int] = mapped_column(primary_key=True)
+    name: Mapped[str] = mapped_column(String(32))
