+++ backend/models/user.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+__authors__ = ["David Foss", "Gabrian Chua", "Nolan Welch", "Rohan Kashyap"]
+__copyright__ = "Copyright 2025"
+__license__ = "MIT"
+
+
+class User(BaseModel):
+    pass
